Add files via upload
This replaces the buttons to select modes and optimizes the code to reduce harmonics in the handpiece. There are still 3 modes, the variable pedal, the selection on the potentiometers and the pulses mode (On for half second and off for half second). The digital button 2, when pressed, selects the variable pedal mode. The digital button 3, when pressed, selects the potentiometer selection mode. The digital button 4, when pressed, selects the pulses mode. The correction value for the variable pedal is the duty potentiometer (Analog input on pin A2). 
If no buttons or potentiometers are used here please read lines 239 and 240. The correction value should be tested. My opinion is that the bare minimun for the machine to work are 1 pedal and 1 potentiometer (The correction potentiometer)